add starsurge lunar eclipse crit bonus
